// src/common/enums/candidate-status.enum.ts
export enum CandidateStatusEnum {
  APPLIED = 'APPLIED',

  SCREENING = 'SCREENING',

  ASSESSMENT_SCHEDULED = 'ASSESSMENT_SCHEDULED',

  ASSESSMENT_CLEARED = 'ASSESSMENT_CLEARED',

  TECHNICAL_SCHEDULED = 'TECHNICAL_SCHEDULED',

  TECHNICAL_CLEARED = 'TECHNICAL_CLEARED',

  HR_SCHEDULED = 'HR_SCHEDULED',

  INTERVIEW_SCHEDULED = 'INTERVIEW_SCHEDULED',

  INTERVIEWED = 'INTERVIEWED',

  SELECTED = 'SELECTED',

  REJECTED = 'REJECTED',

  HIRED = 'HIRED',
}